+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 1.2.0
 - :star2: Feature #14 - Added possibility to create subchains.
+- :star2: Feature #34 - Allow chain's to be observed to see progress.
+- :star2: Feature #35 - Added a symfony console output to display nice output based on the chain's progress.
 
 # 1.1.4
 - :star2: Split item operation can now return more complex data sets using the `keyName` option